Migrate global state to redux toolkit
