--- src/global/globalActions.ts
@@ -1,45 +1,0 @@
-/**
- * See the NOTICE file distributed with this work for additional information
- * regarding copyright ownership.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { createAction } from 'typesafe-actions';
-import { ActionCreator, Action } from 'redux';
-import { ThunkAction } from 'redux-thunk';
-
-import { BreakpointWidth, ScrollPosition } from './globalConfig';
-import { getBreakpointWidth } from './globalSelectors';
-
-import { RootState } from 'src/store';
-
-export const setBreakpointWidth = createAction(
-  'browser/update-breakpoint-width'
-)<BreakpointWidth>();
-
-export const saveScrollPosition = createAction(
-  'update-scroll-position'
-)<ScrollPosition>();
-
-export const updateBreakpointWidth: ActionCreator<
-  ThunkAction<void, any, null, Action<string>>
-> =
-  (viewportWidth: BreakpointWidth) =>
-  async (dispatch, getState: () => RootState) => {
-    const state = getState();
-    const currentBreakpointWidth = getBreakpointWidth(state);
-
-    if (viewportWidth !== currentBreakpointWidth) {
-      dispatch(setBreakpointWidth(viewportWidth));
-    }
-  };
